Move error pages into error folder
