+++ src/components/VideoPage.js
@@ -12,8 +12,8 @@ import { getTranscript } from '../actions/transcript.js';
 import Layout from './Layout';
 import Transcript from './Transcript'
 import Loading from './Loading.js';
-import NotFound from './NotFound.js'
-import InternalError from './InternalError.js'
+import NotFound from './errors/NotFound.js'
+import InternalError from './errors/InternalError.js'
 import PiazzaQuestionForm from './PiazzaQuestionForm.js'
 
 import './VideoPage.css';
