+++ src/components/errors/InternalError.js
@@ -1,17 +1,18 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
-import Layout from './Layout.js';
+import Layout from '../Layout.js';
 
-import './Forbidden.css';
+import './InternalError.css';
 
-export default class Forbidden extends Component {
+export default class InternalError extends Component {
   render() {
     return (
       <Layout>
-        <div className='container container-forbidden'>
+        <div className='container container-not-found'>
           <div>
-            <h1>🔒 You don't seem to have access to this page</h1>
+            <h1>🛸 We're out of service</h1>
+            <p>Please retry your request in a few minutes</p>
             <p><Link to='/'><span className='fa fa-arrow-left' /> Follow here to go back to home page</Link></p>
           </div>
         </div>
