+++ src/components/errors/NotFound.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
-import Layout from './Layout.js';
+import Layout from '../Layout.js';
 
 import './NotFound.css';
